requirements.txt
asgiref==3.5.2
Django==4.1
gunicorn==20.1.0
Pillow==9.2.0
sqlparse==0.4.2
pytz==2020.1
whitenoise==6.2.0
djangorestframework==3.14.0
django-cors-headers==3.13.0